Add files via upload
In this project, text mining and sentiment analysis were applied to a dataset comprising diverse reviews of products and services from three different online brands. The 
analysis aims to score and classify the reviews as either negative, positive, or neutral, using natural language processing. Based on the classified reviews, the task takes further attempts to build a classification model that can take instant custom reviews as input and analyse the sentiments.